+++ src/components/inventory/inventory.html
@@ -1,0 +1,9 @@
+<div id="inventory" #inventory>
+    <div class="inv-grid">
+        <div class="row">
+            <div class="cell">
+
+            </div>
+        </div>
+    </div>
+</div>
