+++ src/components/inventory/inventory.scss
@@ -1,0 +1,36 @@
+.inv-grid{
+    background-color: #575757;
+    display: inline-block;
+    padding: 3px;
+    margin: 5px;
+
+    .row{
+        font-size: 0px;
+
+        .cell{
+            margin: 1px;
+            font-size: initial;
+            display: inline-block;
+            width: 30px;
+            height: 30px;
+            background-color: #2c2c2c;
+
+            &:first-child {
+                border-left-width: 2px;
+            }
+
+            &:last-child{
+                border-right-width: 2px;
+            }
+
+            &[filled="true"]{
+                background-color: red;
+            }
+
+            .item{
+                background-color: #cecece;
+                position: absolute;
+            }
+        }
+    }
+}
