+++ src/components/inventory/item.ts
@@ -1,0 +1,6 @@
+export function Item () {
+
+    this.init = (options) => {
+        Object.keys(options).forEach((key)=>{this[key] = options[key]});
+    }
+}
