+++ src/pages/home/game/game.html
@@ -3,14 +3,7 @@
 
     </div>
     <div id="menuContainer">
-        <div id="inventory" #inventory>
-            <div class="inv-grid">
-                <div class="row">
-                    <div class="cell">
-
-                    </div>
-                </div>
-            </div>
-        </div>
+        <inventory rows="1" columns="4" id="pockets"></inventory>
+        <inventory rows="5" columns="6"></inventory>
     </div>
 </ion-content>
