+++ src/providers/Inventories.ts
@@ -1,0 +1,14 @@
+import { Injectable } from "@angular/core";
+
+@Injectable()
+export class Inventories {
+    private list: any = {};
+    
+    constructor() {
+
+    }
+
+    submitInventory(inventory) {
+        
+    }
+}
